+++ tests/test_frozen_at_when_copying_dataset.py
@@ -18,12 +18,12 @@ def test_frozen_at_value_when_copying_dataset(tmp_dir_fixture):  # NOQA
     os.mkdir(dest_base_uri)
 
     src_dataset = dc.DataSet.from_uri(src_uri)
-    src_frozen_at = src_dataset._admin_metadata["frozen_at"]
+    src_frozen_at = src_dataset.admin_metadata["frozen_at"]
 
     time.sleep(0.1)
 
     dest_uri = dc.copy(src_uri, dest_base_uri)
     dest_dataset = dc.DataSet.from_uri(dest_uri)
-    dest_frozen_at = dest_dataset._admin_metadata["frozen_at"]
+    dest_frozen_at = dest_dataset.admin_metadata["frozen_at"]
 
     assert src_frozen_at == dest_frozen_at